requirements.txt
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
joblib>=1.3
networkx>=3.2
vmdpy>=0.2
PyYAML>=6.0
scikit-fuzzy>=0.4

